Add  endings
Activating the bomb with timer triggers a endGameFact which is used to activate a hole to leave on floor 1 and activates perma death
