--- Descension/Assets/Scripts/Actor/Player/MovePlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d0ade3f4138b5ed4ea1f8153de1c6835
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
