--- Descension/Assets/Scripts/Environment/ClimbHole.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d811d12e36308e74dbc86bfbc9746dd4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
