+++ Descension/Assets/Scripts/Environment/MoveLocation.cs
@@ -1,0 +1,23 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Managers;
+using Util.Enums;
+
+namespace Environment
+{
+    public class MoveLocation : MonoBehaviour
+    {
+        public FactKey Fact;
+        public int x;
+        public int y;
+        // Start is called before the first frame update
+        void Start()
+        {
+            if (FactManager.IsFactTrue(Fact))
+            {
+                transform.position = new Vector3(x, y, 0);
+            }
+        }
+    }
+}
